+++ preprocess.lua
@@ -22,6 +22,10 @@ local options = {
     {
       valid = onmt.utils.ExtendedCmdLine.nonEmpty
     }
+  },
+  {
+    '-check_plength', false,
+    [[Check source and target have same length (for seq tagging).]]
   }
 }
 
@@ -48,6 +52,16 @@ local function isValid(seq, maxSeqLength)
     return seq:size(1) > 0 and seq:size(1) <= maxSeqLength
   end
   return #seq > 0 and #seq <= maxSeqLength
+end
+
+local function parallelCheck(idx, _, _, tokens)
+  local length1 = (type(tokens[1])=='table' and #tokens[1]) or (tokens[1]:dim()==0 and 0) or tokens[1]:size(1)
+  local length2 = (type(tokens[2])=='table' and #tokens[2]) or (tokens[2]:dim()==0 and 0) or tokens[2]:size(1)
+  if length1~=length2 then
+    _G.logger:warning('SENT %s: source/target not aligned (%d/%d)', tostring(idx), length1, length2)
+    return false
+  end
+  return true
 end
 
 local function main()
@@ -96,19 +110,25 @@ local function main()
   end
 
   _G.logger:info('Preparing training data...')
+
+  local parallelValidFunc = nil
+  if opt.check_plength then
+    parallelValidFunc = parallelCheck
+  end
+
   data.train = {}
   if dataType == 'monotext' then
     data.train.src = Preprocessor:makeMonolingualData(opt.train, data.dicts.src, isValid)
   elseif dataType == 'feattext' then
     data.train.src, data.train.tgt = Preprocessor:makeFeatTextData(opt.train_src, opt.train_tgt,
                                                                    data.dicts.tgt,
-                                                                   isValid)
+                                                                   isValid, parallelCheck)
     -- record the size of the input layer
     data.dicts.srcInputSize = data.train.src.vectors[1]:size(2)
   else
     data.train.src, data.train.tgt = Preprocessor:makeBilingualData(opt.train_src, opt.train_tgt,
                                                                     data.dicts.src, data.dicts.tgt,
-                                                                    isValid)
+                                                                    isValid, parallelCheck)
   end
 
   _G.logger:info('')