+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix importance sampling when using multiple GPUs
 * Fix language models training
 * SeqTagger supports input vector too
+* New option in preprocess to check that sizes of source and target are equal (for seqtagging)
 
 ## [v0.7.0](https://github.com/OpenNMT/OpenNMT/releases/tag/v0.7.0) (2017-05-19)